# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SeanAI Assistance API",
        "description": "Speech practice and AI-powered analysis API for SpeechFlow. Provides authentication, challenge browsing, practice session management, and AI-driven speech recording analysis.",
        "contact": {
            "name": "API Support",
            "email": "support@speechflow.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get current user profile",
                "description": "Returns the authenticated user's profile including computed statistics (sessions count, total minutes, average score) and achievements.",
                "operationId": "c6a79007a56511faa6f48397a2d15867",
                "responses": {
                    "200": {
                        "description": "Profile with statistics and achievements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "user@example.com"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Aspiring public speaker",
                                                    "nullable": true
                                                },
                                                "profile_picture": {
                                                    "type": "string",
                                                    "example": "/storage/profile_pictures/abc123.jpg",
                                                    "nullable": true
                                                },
                                                "speaking_goals": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "build_confidence",
                                                        "improve_clarity"
                                                    ]
                                                },
                                                "notification_preferences": {
                                                    "properties": {
                                                        "daily_reminder": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "weekly_report": {
                                                            "type": "boolean",
                                                            "example": false
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "statistics": {
                                                    "properties": {
                                                        "sessions": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "minutes": {
                                                            "type": "integer",
                                                            "example": 45
                                                        },
                                                        "avg_score": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 7.8
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "achievements": {
                                                    "properties": {
                                                        "first_session": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "five_sessions": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "ten_sessions": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "pro_speaker": {
                                                            "type": "boolean",
                                                            "example": false
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile",
                "description": "Updates the authenticated user's profile. Supports partial updates — only send the fields you want to change. Profile picture is uploaded as multipart form data.",
                "operationId": "42803079b1b23a321f57eae40c98b303",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Aspiring public speaker"
                                    },
                                    "profile_picture": {
                                        "description": "Image file (JPEG, PNG, JPG). Max 2 MB.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "speaking_goals": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "build_confidence",
                                            "improve_clarity"
                                        ]
                                    },
                                    "notification_preferences[daily_reminder]": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "notification_preferences[weekly_report]": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "Aspiring public speaker"
                                                },
                                                "profile_picture": {
                                                    "type": "string",
                                                    "example": "/storage/profile_pictures/abc123.jpg"
                                                },
                                                "speaking_goals": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "notification_preferences": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete account",
                "description": "Permanently deletes the authenticated user's account after verifying their current password. All tokens are revoked immediately.",
                "operationId": "f8574ba4a5f3de43a0f6fa1d7181b8e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "Current account password for confirmation",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Account deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Incorrect password"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/profile/history": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get practice history",
                "description": "Returns all practice sessions for the authenticated user with their challenges and results, ordered by most recent first.",
                "operationId": "d611ee5e70447bc1423d0f15c716e3bb",
                "responses": {
                    "200": {
                        "description": "List of practice sessions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PracticeSession"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/profile/share-progress": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get shareable progress summary",
                "description": "Returns a text summary of the user's progress suitable for sharing on social media or messaging.",
                "operationId": "bc398cc60ca295802024d39d52dc7a75",
                "responses": {
                    "200": {
                        "description": "Shareable progress summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "summary": {
                                            "type": "string",
                                            "example": "John's SpeechFlow Progress:\nSessions: 12\nTotal Minutes: 45\nAvg Score: 7.8"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "public-speaking"
                    },
                    "name": {
                        "type": "string",
                        "example": "Public Speaking"
                    },
                    "description": {
                        "type": "string",
                        "example": "Practice your public speaking skills"
                    },
                    "icon": {
                        "type": "string",
                        "example": "microphone"
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Challenge": {
                "title": "Challenge",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduce yourself to a new team"
                    },
                    "suggested_time_minutes": {
                        "type": "integer",
                        "example": 5
                    },
                    "hints_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "tips": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Speak clearly",
                            "Make eye contact"
                        ]
                    },
                    "category": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Category"
                            }
                        ],
                        "nullable": true
                    },
                    "level": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Level"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PracticeSession": {
                "title": "PracticeSession",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Introduce yourself Practice - Feb 08, 2026 10:30 AM"
                    },
                    "challenge": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Challenge"
                            }
                        ],
                        "nullable": true
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "started",
                            "recorded",
                            "processed",
                            "failed"
                        ],
                        "example": "processed"
                    },
                    "result": {
                        "description": "Included when the result relationship is loaded",
                        "properties": {
                            "transcription": {
                                "type": "string"
                            },
                            "feedback": {
                                "type": "string"
                            },
                            "score": {
                                "type": "number",
                                "format": "float"
                            },
                            "metadata": {
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the Bearer token obtained from /auth/login or /auth/register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}